requirements.txt
Django
South
psycopg2
djangorestframework
fabric
gunicorn
